Store svg element attributes (height, width, x, y and angle) in service
Service manages a list of objects (and object for each element) and updates the view accordingly by setting the list to a local scope object on the controller.
These are rudimentary and can be optimized/improved upon.
Limit the draggable area such that the svg cannot be dragged off of the screen.